Modify content generation approach to prevent page refresh
Rather than manually changing the URL to include the updated query string, prompting a page refresh, utilise the document history API to silently update the query string and reload the appropriate index content on click of a pagination link. If the link is copied and pasted to include a page value, the index (and pagination) content are also built accordingly.