Modify 'load more' content generation to prevent page refresh
As with the content generation in the page-based pagination, the document history API can be utilised to update the query string without prompting a page refresh, appending the latest elements to the content rather than rebuilding the document entirely. If the link is copied and pasted to include a page value, the index (and pagination) content are also built accordingly.