Author DELETE function to delete existing JSON documents
Creating a JSON document with POST generates a UUID that is utilised as the file name, serving as a unique identifier of the file itself. A DELETE request tries to delete an existing document with a provided unique identifier and returns the following HTTP response status: 404, if the document does not exist; 500, if the document exists but an error occurs in deletion; or 200, if the document exists and is successfully deleted.